Server-to-device `invoicePaid` messages are acked
Any time an invoice is paid, the coordinator saves a record to a
database collection (DeviceUnackedSettledInvoiceManager). It then sends
a message to the device. The device will ack the message, and
DeviceUnackedSettledInvoiceManager will delete the record. If the device
is offline or otherwise unreachable when the invoice is paid, the
coordinator will send/resend the message when the device comes back
online.