Device UI caches latest device version
It can then use it as a fallback if it is offline or otherwise unable to
load the data and inventory. We're also cleaning up how
`deviceDataManager` is used. Previously the `loaded` state was used when
an existing device was loaded from the server, and the `error` state was
used if the device didn't exist. Now the `loaded` state is used in both
cases, the data being `null` in the latter case. The `error` state is
used when the request times out or otherwise fails.