Devices can invoke a device data fetch
This inverts the initial fetch of device data. The device can invoke a
fetch of its data, rather than the server pushing data to the device
upon socket connection. When the device relied on the server to push
data, it would have no way to timeout and give up waiting for data. This
meant it would wait forever for data that might never come. Now the
device can timeout and give up waiting for data, displaying an error
message to the user instead of showing a never-ending loading spinner.